Uniquify certficate subject names by adding a timestamp.
This fixes a bug where reinstalling an rBuilder results in a
non-bypassable browser error (in firefox, if not others) because the
server certificate has the same issuer name and the same serial number
but a different fingerprint and does not validate. It is illegal in
principle to have two CAs with the same subject name for this very
reason.