Close Mingle #534, restrict queryset tagging strange loops & self awareness
Since we have only one tagged_date field for a queryset, we can't use the
tagged date to decide if a result is stale, because we might just be asking
for a PART of it, and haven't looked at the other part.  This change makes
the querysets use the tags for caching only when getting at the 'all'
collection, which is the most important, leaving the other heads of the
collection to be manually evaluated if accessed directly.  Accessing and
tagging a 'head' of a queryset, like chosen/child/etc, should never
invalidate the 'all' tags or update the tagged date of the queryset itself.
Ideally we'd keep seperate dates for each head, but for now, we have the one.