CreateResource (new permission) is now required to create projects.  Ideally it will be used to gate access to lots of other types of resources, but ahead of delegation (Q1 2012 or later 2011) some things remain admin only, also systems must be anonymously createable for now.
Future commits will auto add created resources to where their CreateResource
points, for use with the "My Querysets" feature.

Automation should update to make sure the additional permissions are granted,
and also test the negative case.

CreateResource is not required on a PBS because those are created as a side
effect and not by an explicit call.  Branch creation looks towards the P.