My Querysets are automatically added when a user is created with a can_create bit true, or the bit is toggled on when editing the user.  When toggled off, they are removed.  An identity role and appropriate grants are created at the same time.
So far these are added for projects & systems, images pending unification.

We still want to hide identity querysets (and likely roles) from the admin
user to avoid spamming them with too many items to manage, and we still have
to do the "add queryset to chosen" for the querysets.   May also want to
change the dynamic nature of them to use chosen querysets instead, TBD.

Also removed some hacks in the users module to make saving/editing users
cleaner.