Killing the very imcomplete changelogs service (minus the table removal, which we can do in P4, as I don't want to invalidate migration testing).
The "changelogs" module was an idea that logged edits to only a few
certain fields, namely systems, in the database.   The database should not be used
for logging, and user info should be annotated.  This can't meet Sarbox
style requirements, nor can it be aggregrated and associated easily
with other events on the system.  Exposing the log history via a paged
collection is probably also not a great idea.  Basically logging is a very
good idea, but we need a better design for it, and it should not be based
around edits to fields, nor should it require annotating certain fields
in the model.  I can see something done at view level in the future that
keeps user information, though we should also consider what *else* we
are logging and how to keep that unified with the other logs in our system.