Flesh out target filter descriptors by adding shortnames to the models.
Also hide any discovered relations without comments, as this most likely means
this is something we want to search by.  The XML API will still allow
searching on anything, but this keeps from having python looking
variables in the dropdowns -- and searching by ID in the GUI
is generally not a good practice.