A workaround that prevents "Stage name" from being eclipsed by "Project name" when searching a Stage because the default search key for Stage is actually project name.
This is accounted for in the filter descriptor so no UI change is required,
but basically the string "literal:" is prefixed before the search field
to keep "name" from being thought to imply something else.  This is doable
on any search key but is not something we should have to do very frequently.