Disable various sketchy things for demo.  Namely, we are saving desired_top_level_items in two (2!!) different formats, the shorter one, various conary hooks that generate the config descriptor do not like.  So this means you must take a new survey before configuring, for now, and NOT rely on the survey you get when you initially register.
Additionally, this means for various safety reasons the conary cache in mint
is still disabled, though I had some success in enabling it earlier today, if
any systems still have the "shorter" trove specs this would cause them to
choke, which we do not want.

When we go back in and fix this, and everything saves the same way, the system
updates flag also needs to consider what project something is in, which it
presently does not do.

Hopefully the long trovespecs are already in the survey, if they are not, this
may be something both windows and CIM clients may need to add to keep
everything running.