Let rmake3 push status up to the API instead of polling it
Polling from the API is both expensive and prone to race conditions
because rmake would push the results up at the end of the job. If a poll
was in progress at the same time, it could end up overwriting the job
results, leaving the job finished but without any results.